feat!: replace old codes with new pattern
