+++ src/contracts/index.ts
@@ -1,7 +1,6 @@
-import * as Project from './Project-clone.js';
-import * as Campaign from './Campaign-clone.js';
-import * as Participation from './Participation-clone.js';
-import * as Funding from './Funding-clone.js';
-// import * as Treasury from './Treasury.js';
+import * as Project from './Project.js';
+import * as Campaign from './Campaign.js';
+import * as Participation from './Participation.js';
+import * as Funding from './Funding.js';
 
 export { Project, Campaign, Participation, Funding };
