+++ src/index.ts
@@ -1,90 +1,3 @@
-// export * as Constants from './Constants.js';
-// export * as Storage from './contracts/storages.js';
-// export * as ZkApp from './contracts/index.js';
-
-// export {
-//     // Structs
-//     CheckProjectOwnerInput,
-//     CreateProjectInput,
-//     UpdateProjectInput,
-//     CreateProjectProofOutput,
-
-//     // Zk Programs & Proofs
-//     CreateProject,
-//     ProjectProof,
-
-//     // Smart Contract
-//     ProjectContract,
-
-//     // Actions & Events
-//     ProjectAction,
-//     EventEnum as ProjectEvents,
-// } from './contracts/Project.js';
-
-// export {
-//     // Smart Contract
-//     CampaignContract,
-
-//     // Actions & Events
-//     CampaignAction,
-// } from './contracts/Campaign-clone.js';
-
-// export {
-//     // Structs
-//     JoinCampaignInput,
-//     CheckParticipationIndexInput,
-//     CreateParticipationProofOutput,
-
-//     // Zk Programs & Proofs
-//     JoinCampaign,
-//     ParticipationProof,
-
-//     // Smart Contract
-//     ParticipationContract,
-
-//     // Actions & Events
-//     ParticipationAction,
-//     EventEnum as ParticipationEvents,
-// } from './contracts/Participation.js';
-
-// export {
-//     // Structs
-//     RequestSent,
-//     CustomScalarArray,
-//     FundingInput,
-//     CheckValueInput,
-//     ReduceOutput as ReduceFundingOutput,
-//     RollupActionsOutput as RollupParticipationActionsOutput,
-
-//     // Zk Programs & Proofs
-//     ReduceProof,
-//     CreateReduceProof,
-//     ProofRollupAction,
-//     CreateRollupProof,
-
-//     // Smart Contract
-//     FundingContract,
-
-//     // Actions & Events
-//     FundingAction,
-//     EventEnum as FundingEvents,
-// } from './contracts/Funding.js';
-
-// export {
-//     // Structs
-//     InvestVector,
-//     ClaimFundInput,
-//     CheckIfNotClaimedInput,
-//     ClaimFundProofOutput,
-
-//     // Zk Programs & Proofs
-//     TreasuryProof,
-//     ClaimFund,
-
-//     // Smart Contract
-//     TreasuryContract,
-
-//     // Actions & Events
-//     TreasuryAction,
-//     EventEnum as TreasuryEvents,
-// } from './contracts/Treasury.js';
+export * as Storage from './storages/index.js';
+export * as ZkApp from './contracts/index.js';
+export * as Constants from './Constants.js';
