+++ src/storages/SharedStorage.ts
@@ -9,7 +9,8 @@ import {
     Struct,
 } from 'o1js';
 import { ErrorEnum, INSTANCE_LIMITS, ZkAppEnum } from '../Constants.js';
-import { buildAssertMessage } from '../libs/utils.js';
+import { Utils } from '@auxo-dev/auxo-libs';
+// import { buildAssertMessage } from '../libs/utils.js';
 
 export const ZKAPP_ADDRESS_TREE_HEIGHT =
     Math.ceil(Math.log2(INSTANCE_LIMITS.ZKAPP_ADDRESS_TREE_SIZE)) + 1;
@@ -115,61 +116,6 @@ export const enum ActionStatus {
     REDUCED,
 }
 
-export class ReduceStorage {
-    private _actionMap: MerkleMap;
-    private _actions: { [key: string]: Field };
-
-    constructor(actions?: { actionState: Field; status: ActionStatus }[]) {
-        this._actionMap = EMPTY_REDUCE_MT();
-        this._actions = {};
-        if (actions) {
-            for (let i = 0; i < actions.length; i++) {
-                this.updateLeaf(
-                    actions[i].actionState,
-                    ReduceStorage.calculateLeaf(actions[i].status)
-                );
-            }
-        }
-    }
-
-    get root(): Field {
-        return this._actionMap.getRoot();
-    }
-
-    get actionMap(): MerkleMap {
-        return this._actionMap;
-    }
-
-    get actions(): { [key: string]: Field } {
-        return this._actions;
-    }
-
-    static calculateLeaf(status: ActionStatus): Field {
-        return Field(status);
-    }
-
-    calculateLeaf(status: ActionStatus): Field {
-        return ReduceStorage.calculateLeaf(status);
-    }
-
-    static calculateIndex(actionState: Field): Field {
-        return actionState;
-    }
-
-    calculateIndex(actionState: Field): Field {
-        return ReduceStorage.calculateIndex(actionState);
-    }
-
-    getWitness(index: Field): MerkleMapWitness {
-        return this._actionMap.getWitness(index);
-    }
-
-    updateLeaf(index: Field, leaf: Field): void {
-        this._actionMap.set(index, leaf);
-        this._actions[index.toString()] = leaf;
-    }
-}
-
 export function getZkAppRef(
     map: AddressMT,
     index: ZkAppEnum | number,
@@ -196,11 +142,19 @@ export function verifyZkApp(
 ) {
     root.assertEquals(
         ref.witness.calculateRoot(Poseidon.hash(ref.address.toFields())),
-        buildAssertMessage(programName, 'verifyZkApp', ErrorEnum.ZKAPP_ROOT)
+        Utils.buildAssertMessage(
+            programName,
+            'verifyZkApp',
+            ErrorEnum.ZKAPP_ROOT
+        )
     );
 
     key.assertEquals(
         ref.witness.calculateIndex(),
-        buildAssertMessage(programName, 'verifyZkApp', ErrorEnum.ZKAPP_INDEX)
+        Utils.buildAssertMessage(
+            programName,
+            'verifyZkApp',
+            ErrorEnum.ZKAPP_INDEX
+        )
     );
 }
