+++ src/storages/index.ts
@@ -1,8 +1,7 @@
 import * as CampaignStorage from './CampaignStorage.js';
-import * as ProjectStorage from '../contracts/ProjectStorage.js';
+import * as ProjectStorage from './ProjectStorage.js';
 import * as ParticipationStorage from './ParticipationStorage.js';
 import * as FundingStorage from './FundingStorage.js';
-import * as TreasuryStorage from './TreasuryStorage.js';
 import * as SharedStorage from './SharedStorage.js';
 
 export {
@@ -10,6 +9,5 @@ export {
     ProjectStorage,
     ParticipationStorage,
     FundingStorage,
-    TreasuryStorage,
     SharedStorage,
 };
